Start of chapter 7 (more notes in description)
Completed draft of "Visual Communiation...".  Started "Cognitive
Science...". Images used so far are placeholders to see if we can get
Wiley ones, except for the longitude chart which I'd like to make sure
stays in the book, but Wiley may have a better source image for it.
Going to see if I can write some R code for the various examples vs
just crank up photoshop. That prbly means making a source code folder
for the book sooner than later.