changed figure; created folder for figures
I though it might be more interesting to let a figure continue the
example of ever increasingly complex visualizations evolving over time.
def want to discuss.

we'll also prbly need to have actual folders under the figures folder
(per-chapter)

and, i'll need to re-read the "how to size figures" instructions which
must be somewhere in our array of Wiley commandments.